Nicer grid
Implemented a nicer-looking grid with the use of underlining and unicode vertical half-blocks.
A putty screenshot is attached. The font used is Consolas.
Does not seem to work well on real framebuffer console.
Give a compile warning on Ubuntu 18.04, but seems to be working. Couldn't figure out why the compiler complains.
Compiles on my slackware-current with no warnings.